+++ src/core/helpers/deviceTools.spec.js
