+++ src/core/plugins/fastboot/plugin.spec.js
@@ -1,7 +1,7 @@
 const mainEvent = { emit: jest.fn() };
 beforeEach(() => mainEvent.emit.mockReset());
 
-const { fastboot } = require("../../deviceTools.js");
+const { fastboot } = require("../../helpers/deviceTools.js");
 const fastbootPlugin = new (require("./plugin.js"))({}, "a", mainEvent);
 
 describe("fastboot plugin", () => {
